project1 tests passing

+++ inc/Common/conversion.h
@@ -23,7 +23,7 @@ int32_t my_atoi(uint8_t *ptr, uint8_t digits, uint32_t base);
 */
 
 
-int32_t my_itoa(int32_t data, uint8_t *ptr, uint32_t base);
+uint8_t my_itoa(int32_t data, uint8_t *ptr, uint32_t base);
 
 /*
 @description:function to convert 32 bit integer data to ascii string and copy to the pointer address
